+++ dataclass_array/utils/tree_utils_test.py
@@ -1,4 +1,4 @@
-# Copyright 2023 The dataclass_array Authors.
+# Copyright 2024 The dataclass_array Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.